move 'forEach' tests to separate file
