+++ types/ramda/tsconfig.json
@@ -82,6 +82,7 @@
         "test/findLast-tests.ts",
         "test/findLastIndex-tests.ts",
         "test/flatten-tests.ts",
-        "test/flip-tests.ts"
+        "test/flip-tests.ts",
+        "test/forEach-tests.ts"
     ]
 }
